+++ edc-c-client/org.eclipse.paho.mqtt.c/src/utf-8.c
@@ -110,7 +110,7 @@ int UTF8_validateString(char* string)
 	int rc = 0;
 
 	FUNC_ENTRY;
-	rc = UTF8_validate((int)(strlen(string)), string);
+	rc = UTF8_validate(strlen(string), string);
 	FUNC_EXIT_RC(rc);
 	return rc;
 }
